+++ test/src/helpers/config.ts
@@ -1,0 +1,2 @@
+export const specs = 'specs/spec/**/*.spec.js';
+export const components = 'specs/components/**/*.iif.js';